refactor: move and extract UnitList, UnitListItem and UnitListItemLessonCount
