+++ src/components/TeacherComponents/UnitListItem/UnitListItem.stories.tsx
@@ -1,9 +1,10 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import AnalyticsDecorator from "../../../../storybook-decorators/AnalyticsDecorator";
-import unitListingFixture from "../../../../node-lib/curriculum-api/fixtures/unitListing.fixture";
-
 import UnitListItem, { UnitListItemProps } from "./UnitListItem";
+
+import AnalyticsDecorator from "@/storybook-decorators/AnalyticsDecorator";
+import unitListingFixture from "@/node-lib/curriculum-api/fixtures/unitListing.fixture";
+
 
 const unitListing = unitListingFixture();
 //@ts-expect-error : propData is not assignable to type 'UnitListItemProps' on account of ? in type
