+++ src/components/UnitAndLessonLists/ListItemHeader.tsx
@@ -2,10 +2,10 @@ import { FC, MutableRefObject } from "react";
 
 import OakLink from "../OakLink";
 import { LessonListItemProps } from "../TeacherComponents/LessonListItem";
+import { UnitListItemProps } from "../TeacherComponents/UnitListItem/UnitListItem";
 
 import CategoryHeading from "./CategoryHeading";
 import ExemplarCategoryHeading from "./ExemplarCategoryHeading";
-import { UnitListItemProps } from "./UnitList/UnitListItem/UnitListItem";
 
 import { Heading } from "@/components/SharedComponents/Typography";
 import Flex from "@/components/SharedComponents/Flex";
